requirements.txt
# pip run -r requirements.txt

fastapi 
uvicorn
requests

streamlit

# chainlit
openai
python-dotenv
-e .